updated views per MVP
