+++ components/forms/MemberForm.js
@@ -34,13 +34,13 @@ export default function MemberForm({ memberObj }) {
   const handleSubmit = (e) => {
     e.preventDefault();
     if (memberObj.firebaseKey) {
-      updateMember(formInput).then(() => router.push('/team'));
+      updateMember(formInput).then(() => router.push('/'));
     } else {
       const payload = { ...formInput, uid: user.uid };
       createMember(payload).then(({ name }) => {
         const patchPayload = { firebaseKey: name };
         updateMember(patchPayload).then(() => {
-          router.push('/team');
+          router.push('/');
         });
       });
     }
