+++ pages/index.js
@@ -1,26 +1,31 @@
-import { Button } from 'react-bootstrap';
-import { signOut } from '../utils/auth';
+import { useEffect, useState } from 'react';
 import { useAuth } from '../utils/context/authContext';
 
+import { getMembers } from '../api/memberData';
+import MemberCard from '../components/MemberCard';
+
 function Home() {
+  const [members, setMembers] = useState([]);
+
   const { user } = useAuth();
 
+  const getAllMembers = () => {
+    getMembers(user.uid).then(setMembers);
+  };
+
+  useEffect(() => {
+    getAllMembers();
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   return (
-    <div
-      className="text-center d-flex flex-column justify-content-center align-content-center"
-      style={{
-        height: '90vh',
-        padding: '30px',
-        maxWidth: '400px',
-        margin: '0 auto',
-      }}
-    >
-      <h1>Hello {user.displayName}! </h1>
-      <p>Click the button below to logout!</p>
-      <Button variant="danger" type="button" size="lg" className="copy-btn" onClick={signOut}>
-        Sign Out
-      </Button>
-    </div>
+    <>
+      <div className="d-flex flex-wrap">
+        {members.map((member) => (
+          <MemberCard memberObj={member} key={member.firebaseKey} onUpdate={getAllMembers} />
+        ))}
+      </div>
+    </>
   );
 }
